TRI_scripts/Kyle/requirements.txt
asttokens==3.0.1
decorator==5.2.1
executing==2.2.1
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
ipdb==0.13.13
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.2.1
numpy==2.4.2
opencv-python==4.13.0.90
pandas==3.0.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
protobuf==6.33.5
psutil==7.2.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scipy==1.17.0
six==1.17.0
stack-data==0.6.3
tqdm==4.67.2
traitlets==5.14.3
wcwidth==0.5.3
